requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.4
asgiref==3.2.10
attrs==19.3.0
billiard==3.3.0.23
black==19.10b0
celery==3.1.26.post2
click==7.1.2
dj-database-url==0.5.0
Django==3.1
django-filter==2.3.0
djangorestframework==3.11.1
gunicorn==20.0.4
importlib-metadata==1.7.0
kombu==3.0.37
mccabe==0.6.1
pathspec==0.8.0
psycopg2==2.8.5
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.1
regex==2020.7.14
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
whitenoise==5.2.0
zipp==3.1.0